requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
python-decouple==3.8
requests==2.32.3
psycopg2-binary==2.9.9 